LayoutTests/fast/frames/seamless/seamless-inline-expected.txt

Test that inline seamless iframes 'shrink-wrap' their contents like inline-blocks do.
PASS window.getComputedStyle(iframe1).display is "inline-block"
PASS window.getComputedStyle(iframe2).display is "inline-block"
PASS window.getComputedStyle(iframe1).width is "150px"
PASS window.getComputedStyle(iframe1).height is "50px"
PASS window.getComputedStyle(parent1).height is "50px"
PASS window.getComputedStyle(iframe2).width is "100px"
PASS window.getComputedStyle(iframe2).height is "100px"
PASS window.getComputedStyle(parent2).height is "150px"
